+++ app/views/app/brands.php
@@ -28,6 +28,10 @@ Partials::Header(true, true);
             
             $brands = Functions::Brands()->getAll();
 
+            if (!$brands) {
+                out("No Brands were found.");
+            }
+
             foreach ($brands as $brand) { ?>
                 <tr data-id="<?php out($brand["id"]); ?>">
                     <td><?php out($brand["name"]); ?></td>
@@ -40,8 +44,10 @@ Partials::Header(true, true);
                         ?>
                     </td>
                     <td>
-                        <a class="edit action-button tooltip" data-id="<?php out($brand["id"]); ?>" data-tooltip="<?php Translator::translate("brands.edit"); ?>" href="/app/edit/brand/<?php out($brand["id"]); ?>"><i class="fa-solid fa-pen"></i></a>
-                        <a class="del action-button tooltip" data-id="<?php out($brand["id"]); ?>" data-tooltip="<?php Translator::translate("brands.delete"); ?>" href="javascript:void(0);" onclick="deleteItem('brand', <?php out($brand['id']); ?>);"><i class="fas fa-trash"></i></a>
+                        <?php if (Functions::Users()->getUserID() == $user_id = $brand["user_id"]) { ?>
+                            <a class="edit action-button tooltip" data-id="<?php out($brand["id"]); ?>" data-tooltip="<?php Translator::translate("brands.edit"); ?>" href="/app/edit/brand/<?php out($brand["id"]); ?>"><i class="fa-solid fa-pen"></i></a>
+                            <a class="del action-button tooltip" data-id="<?php out($brand["id"]); ?>" data-tooltip="<?php Translator::translate("brands.delete"); ?>" href="javascript:void(0);" onclick="deleteItem('brand', <?php out($brand['id']); ?>);"><i class="fas fa-trash"></i></a>
+                        <?php } ?>
                     </td>
                 </tr>
